feat: new can go back component
