+++ packages/shared/src/components/post/SquadPostPageNavigation.tsx
@@ -1,13 +1,11 @@
 import React, { ReactElement } from 'react';
 import Link from 'next/link';
-import { useRouter } from 'next/router';
 import ArrowIcon from '../icons/Arrow';
 import { IconSize } from '../Icon';
 import { feature } from '../../lib/featureManagement';
 import { FeedLayout } from '../../lib/featureValues';
-import { ButtonVariant } from '../buttons/common';
-import { Button } from '../buttons/ButtonV2';
 import { useFeature } from '../GrowthBookProvider';
+import { PostBackButton } from './common/PostBackButton';
 
 interface SquadPostPageNavigationProps {
   squadLink: string;
@@ -34,23 +32,11 @@ function SquadPostPageNavigation({
   squadLink,
 }: SquadPostPageNavigationProps): ReactElement {
   const isFromNavigation = checkIsFromNavigation();
-  const router = useRouter();
   const link = isFromNavigation ? NOTIFICATIONS_LINK : squadLink;
   const layout = useFeature(feature.feedLayout);
 
   if (layout === FeedLayout.V1) {
-    const canGoBack = !!globalThis?.window?.history?.length;
-
-    return (
-      <Button
-        className="w-fit"
-        onClick={canGoBack ? router.back : () => router.push(squadLink)}
-        variant={ButtonVariant.Tertiary}
-        icon={<ArrowIcon className="-rotate-90" />}
-      >
-        Back
-      </Button>
-    );
+    return <PostBackButton link={squadLink} />;
   }
 
   return (
