+++ packages/webapp/pages/posts/[id]/index.tsx
@@ -47,13 +47,10 @@ import ArrowIcon from '@dailydotdev/shared/src/components/icons/Arrow';
 import { IconSize } from '@dailydotdev/shared/src/components/Icon';
 import Link from 'next/link';
 import { CollectionPostContent } from '@dailydotdev/shared/src/components/post/collection';
-import {
-  Button,
-  ButtonVariant,
-} from '@dailydotdev/shared/src/components/buttons/ButtonV2';
 import { useFeature } from '@dailydotdev/shared/src/components/GrowthBookProvider';
 import { feature } from '@dailydotdev/shared/src/lib/featureManagement';
 import { FeedLayout } from '@dailydotdev/shared/src/lib/featureValues';
+import { PostBackButton } from '@dailydotdev/shared/src/components/post/common/PostBackButton';
 import { getTemplatedTitle } from '../../../components/layouts/utils';
 import { getLayout as getMainLayout } from '../../../components/layouts/MainLayout';
 
@@ -167,48 +164,22 @@ const PostPage = ({ id, initialData }: Props): ReactElement => {
     const routedFromSquad = router?.query?.squad;
     const squadLink = `/squads/${router.query.squad}`;
 
-    if (layout === FeedLayout.Control) {
-      if (routedFromSquad) {
-        return (
-          <Link href={squadLink}>
-            <a className="flex flex-row items-center font-bold text-theme-label-tertiary typo-callout">
-              <ArrowIcon size={IconSize.Medium} className="mr-2 -rotate-90" />
-              Back to {router.query.n || 'Squad'}
-            </a>
-          </Link>
-        );
-      }
+    if (layout === FeedLayout.V1) {
+      return <PostBackButton link={routedFromSquad ? squadLink : undefined} />;
+    }
 
+    if (!routedFromSquad) {
       return null;
     }
 
-    if (globalThis?.window?.history?.length) {
-      return (
-        <Button
-          className="w-fit"
-          onClick={router.back}
-          variant={ButtonVariant.Tertiary}
-          icon={<ArrowIcon className="-rotate-90" />}
-        >
-          Back
-        </Button>
-      );
-    }
-
-    if (routedFromSquad) {
-      return (
-        <Button
-          className="w-fit"
-          onClick={() => router.push(squadLink)}
-          variant={ButtonVariant.Tertiary}
-          icon={<ArrowIcon className="-rotate-90" />}
-        >
-          Back
-        </Button>
-      );
-    }
-
-    return null;
+    return (
+      <Link href={squadLink}>
+        <a className="flex flex-row items-center font-bold text-theme-label-tertiary typo-callout">
+          <ArrowIcon size={IconSize.Medium} className="mr-2 -rotate-90" />
+          Back to {router.query.n || 'Squad'}
+        </a>
+      </Link>
+    );
   })();
 
   const navigation: Record<PostType, ReactNode> = {
